+++ walking-beans-backend/src/main/java/walking_beans/walking_beans_backend/model/dto/DeliveryIncome.java
@@ -15,11 +15,11 @@ import java.sql.Timestamp;
 public class DeliveryIncome {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int incomeIndex;
+    private long incomeIndex;
 
-    private int riderId;
+    private long riderId;
 
-    private int orderId;
+    private long orderId;
 
     private Timestamp incomeDate;
